ADIOS2.jl

ADIOS2.jl is a Julia interface to the ADIOS2, the Adaptable Input Output System version 2.

Adios functions

ADIOS2.Adios — Type

mutable struct Adios

Holds a C pointer adios2_adios *.

This value is finalized automatically. It can also be explicitly finalized by calling finalize(adios).

ADIOS2.init_serial — Function

adios = init_serial()::Adios

Initialize an Adios struct in a serial, non-MPI application. Doesn’t require a runtime config file.

See also the ADIOS2 documentation.

ADIOS2.declare_io — Function

io = declare_io(adios::Adios, name::AbstractString)::AIO

Declare a new IO handler.

See also the ADIOS2 documentation.

ADIOS2.afinalize — Function

afinalize(adios::Adios)

Finalize the ADIOS context adios. It is usually not necessary to call this function.

Instead of calling this function, one can also call the finalizer via finalize(adios). This finalizer is also called automatically when the Adios object is garbage collected.

See also the ADIOS2 documentation