ADIOS2.jl

ADIOS2.jl is a Julia interface to the ADIOS2, the Adaptable Input Output System version 2.

Types

ADIOS2.Error — Type

@enum Error begin
    error_none
    error_invalid_argument
    error_system_error
    error_runtime_error
    error_exception
end

Error return types for all ADIOS2 functions

Based on the library C++ standardized exceptions. Each error will issue a more detailed description in the standard error output, stderr

ADIOS2.Mode — Type

@enum Mode begin
    mode_undefined
    mode_write
    mode_read
    mode_append
    mode_deferred
    mode_sync
end

Mode specifies for various functions. write, read, append are used for file operations, deferred, sync are used for get and put operations.

ADIOS2.ShapeId — Type

@enum ShapeId begin
    shapeid_unknown
    shapeid_global_value
    shapeid_global_array
    shapeid_joined_array
    shapeid_local_value
    shapeid_local_array
end

Adios functions

ADIOS2.Adios — Type

mutable struct Adios

Holds a C pointer adios2_adios *.

This value is finalized automatically. It can also be explicitly finalized by calling finalize(adios).

ADIOS2.adios_init_mpi — Function

adios = adios_init_mpi(MPI.Comm)
adios::Union{Adios,Nothing}

Starting point for MPI apps. Creates an ADIOS handler. MPI collective and it calls MPI_Comm_dup.

ADIOS2.adios_init_serial — Function

adios = adios_init_serial()
adios::Union{Adios,Nothing}

Initialize an Adios struct in a serial, non-MPI application. Doesn’t require a runtime config file.

See also the ADIOS2 documentation.

ADIOS2.declare_io — Function

io = declare_io(adios::Adios, name::AbstractString)
io::Union{AIO,Nothing}

Declare a new IO handler.

See also the ADIOS2 documentation.

ADIOS2.adios_finalize — Function

err = adios_finalize(adios::Adios)
err::Error

Finalize the ADIOS context adios. It is usually not necessary to call this function.

Instead of calling this function, one can also call the finalizer via finalize(adios). This finalizer is also called automatically when the Adios object is garbage collected.

See also the ADIOS2 documentation

IO functions

ADIOS2.AIO — Type

struct AIO

Holds a C pointer adios2_io *.

ADIOS2.define_variable — Function

variable = define_variable(io::AIO, name::AbstractString, type::Type,
                     shape::Union{Nothing,CartesianIndex}=nothing,
                     start::Union{Nothing,CartesianIndex}=nothing,
                     count::Union{Nothing,CartesianIndex}=nothing,
                     constant_dims::Bool=false)
variable::Union{Nothing,Variable}

Define a variable within io.

Arguments

io: handler that owns the variable

name: unique variable identifier

type: primitive type

ndims: number of dimensions

shape: global dimension

start: local offset

count: local dimension

constant_dims: true: shape, start, count won't change; false: shape, start, count will change after definition

ADIOS2.inquire_variable — Function

variable = inquire_variable(io::AIO, name::AbstractString)
variable::Union{Nothing,Variable}

Retrieve a variable handler within current io handler.

Arguments

io: handler to variable io owner

name: unique variable identifier within io handler

ADIOS2.inquire_all_variables — Function

variables = inquire_all_variables(io::AIO)
variables::Union{Nothing,Vector{Variable}}

Returns an array of variable handlers for all variable present in the io group.

Arguments

io: handler to variables io owner

Base.open — Function

engine = open(io::AIO, name::AbstractString, mode::Mode)
engine::Union{Nothing,Engine}

Open an Engine to start heavy-weight input/output operations.

In MPI version reuses the communicator from adios_init_mpi. MPI Collective function as it calls MPI_Comm_dup.

Arguments

io: engine owner

name: unique engine identifier

mode: mode_write, mode_read, `mode_appendq (not yet supported)

Variable functions

ADIOS2.Variable — Type

struct Variable

Holds a C pointer adios2_variable *.

ADIOS2.name — Function

name = name(variable::Variable)
name::Union{Nothing,String}

Retrieve variable name.

ADIOS2.type — Function

type = type(variable::Variable)
type::Union{Nothing,Type}

Retrieve variable type.

ADIOS2.shapeid — Function

shapeid = shapeid(variable::Variable)
shapeid::Union{Nothing,ShapeId}

Retrieve variable shapeid.

Base.ndims — Function

var_ndims = ndims(variable::Variable)
var_ndims::Union{Nothing,Int}

Retrieve current variable number of dimensions.

ADIOS2.shape — Function

var_shape = shape(variable::Variable)
var_shape::Union{Nothing,CartesianIndex}

Retrieve current variable shape.

ADIOS2.start — Function

var_start = start(variable::Variable)
var_start::Union{Nothing,CartesianIndex}

Retrieve current variable start.

Base.count — Function

var_count = count(variable::Variable)
var_count::Union{Nothing,CartesianIndex}

Retrieve current variable count.

ADIOS2.steps_start — Function

var_steps_start = steps_start(variable::Variable)
var_steps_start::Union{Nothing,Int}

Read API, get available steps start from available steps count (e.g. in a file for a variable).

This returns the absolute first available step, don't use with adios2_set_step_selection as inputs are relative, use 0 instead.

ADIOS2.steps — Function

var_steps = steps(variable::Variable)
var_steps::Union{Nothing,Int}

Read API, get available steps count from available steps count (e.g. in a file for a variable). Not necessarily contiguous.

ADIOS2.selection_size — Function

var_selection_size = selection_size(variable::Variable)
var_selection_size::Union{Nothing,Int}

Return the minimum required allocation (in number of elements of a certain type, not bytes) for the current selection.

Base.minimum — Function

var_min = minimum(variable::Variable)
var_min::Union{Nothing,T}

Read mode only: return the absolute minimum for variable.

Base.maximum — Function

var_max = maximum(variable::Variable)
var_max::Union{Nothing,T}

Read mode only: return the absolute maximum for variable.

Engine functions

ADIOS2.Engine — Type

struct Engine

Holds a C pointer adios2_engine *.