ADIOS2.jl

ADIOS2.jl is a Julia interface to the ADIOS2, the Adaptable Input Output System version 2.

Types

ADIOS2.Error — Type

@enum Error begin
    error_none
    error_invalid_argument
    error_system_error
    error_runtime_error
    error_exception
end

Error codes returned from various functions

ADIOS2.Mode — Type

@enum Mode begin
    mode_undefined
    mode_write
    mode_read
    mode_append
    mode_deferred
    mode_sync
end

Mode specifies for various functions. write, read, append are used for file operations, deferred, sync are used for get and put operations.

Adios functions

ADIOS2.Adios — Type

mutable struct Adios

Holds a C pointer adios2_adios *.

This value is finalized automatically. It can also be explicitly finalized by calling finalize(adios).

ADIOS2.init_mpi — Function

adios = init_mpi(MPI.Comm)
adios::Union{Adios,Nothing}

Starting point for MPI apps. Creates an ADIOS handler. MPI collective and it calls MPI_Comm_dup.

ADIOS2.init_serial — Function

adios = init_serial()
adios::Union{Adios,Nothing}

Initialize an Adios struct in a serial, non-MPI application. Doesn’t require a runtime config file.

See also the ADIOS2 documentation.

ADIOS2.declare_io — Function

io = declare_io(adios::Adios, name::AbstractString)
io::Union{AIO,Nothing}

Declare a new IO handler.

See also the ADIOS2 documentation.

ADIOS2.afinalize — Function

err = afinalize(adios::Adios)
err::Error

Finalize the ADIOS context adios. It is usually not necessary to call this function.

Instead of calling this function, one can also call the finalizer via finalize(adios). This finalizer is also called automatically when the Adios object is garbage collected.

See also the ADIOS2 documentation